Use template variables
