+++ media/css/firefox/welcome.scss
@@ -368,180 +368,166 @@ $image-path: '/media/protocol/img';
     }
 }
 
-// hide alt hero images
-#hero-image,
-#hero-animation {
-  display: none;
-}
-
-
 .welcome-page8.variation-2-image,
 .welcome-page8.variation-3-animation {
-  .mzp-l-content {
-    padding: 24px 52px;
-  }
+    .mzp-l-content {
+        padding: 24px 52px;
+    }
 
-  // hide default hero image
-  .mzp-c-hero.variation-default {
-    display: none;
-  }
+    .mzp-c-hero-title {
+        max-width: $content-md;
+        margin: 0 auto;
+    }
+}
 
-  .mzp-c-hero-title {
-    max-width: $content-md;
-    margin: 0 auto;
-  }
-
-  .image-container {
+.image-container {
     margin: 35px 0 70px 0;
 
     .image-wrapper {
-      display: inline-block;
-      vertical-align: top;
-      margin: 0 50px 0 50px;
-      position: relative;
-
-      p {
-        font-size: 17px;
-        max-width: 250px;
-        margin: 0 auto;
-      }
-
-      img {
+        display: inline-block;
         vertical-align: top;
-        height: 290px;
-        width: auto;
-      }
+        margin: 0 50px 0 50px;
+        position: relative;
 
-      video {
-        vertical-align: top;
-        height: 160px;
-        pointer-events: none;
-      }
+        // This test is not being performed in any RTL locales
+        &:first-of-type {
+            margin: 0 50px 0 0;
+        }
+
+        &:last-child {
+            margin: 0 0 0 50px;
+        }
+
+        p {
+            @include font-size(17px);
+            max-width: 250px;
+            margin: 0 auto;
+        }
+
+        img {
+            vertical-align: top;
+            height: 290px;
+            width: auto;
+        }
+
+        video {
+            vertical-align: top;
+            pointer-events: none;
+        }
     }
 
     #arrow-right,
     #arrow-down {
-      position: absolute;
-
-      img {
-        height: 85px;
-      }
+        position: absolute;
     }
 
     #arrow-right {
-      top: 110px;
-      right: -90px;
+        top: 110px;
+        right: -90px;
+        height: 85px;
     }
 
     #arrow-down {
-      display: none;
+        display: none;
     }
-
-    // This test is not being performed in any RTL locales
-    .image-wrapper:first-child {
-      margin: 0 50px 0 0;
-    }
-
-    .image-wrapper:last-child {
-      margin: 0 0 0 50px;
-    }
-  }
 }
 
-@media (max-width: 1185px) {
-  .welcome-page8.variation-2-image {
-    .image-container {
-      .image-wrapper {
-        vertical-align: top;
-        margin: 0 30px 0 30px;
+@media (max-width: 1190px) {
+    .welcome-page8.variation-2-image {
+        .image-container {
+            .image-wrapper {
+                margin: 0 30px 0 30px;
 
-        &:first-child {
-          margin: 0 30px 0 0;
+                &:first-of-type {
+                    margin: 0 30px 0 0;
+                }
+
+                &:last-child {
+                    margin: 0 0 0 30px;
+                }
+
+                p {
+                    @include font-size(16px);
+                    max-width: 178px;
+                }
+
+                img {
+                    height: 178px;
+                }
+
+                #arrow-right {
+                    top: 62px;
+                    right: -55px;
+                    height: 60px;
+                }
+            }
         }
-
-        &:last-child {
-          margin: 0 0 0 30px;
-        }
-
-        p {
-          font-size: 16px;
-          max-width: 178px;
-        }
-
-        img {
-          height: 178px;
-        }
-
-        #arrow-right {
-          top: 62px;
-          right: -55px;
-
-          img {
-            height: 60px;
-          }
-        }
-      }
-
-      @media (max-width: 767px) {
-        max-width: 300px;
-        margin: 5px auto 70px auto;
-
-        #arrow-right {
-          display: none;
-        }
-
-        .image-wrapper {
-          margin: 0;
-
-          &:last-child {
-            margin: 0;
-          }
-
-          &:first-child {
-            margin: 0 0 50px 0;
-          }
-
-          img {
-            height: 250px;
-          }
-
-          p {
-            max-width: 230px;
-          }
-        }
-
-        #arrow-down {
-          display: block;
-          bottom: -63px;
-          left: 84px;
-          display: block;
-
-          img {
-            height: 45px;
-          }
-        }
-      }
     }
-  }
 }
-
-
 
 .welcome-page8.variation-2-image {
-  #hero-image {
-    display: block;
-  }
+    .image-container {
+        @media (max-width: 767px) {
+            max-width: 300px;
+            margin: 5px auto 70px auto;
+
+            #arrow-right {
+                display: none;
+            }
+
+            .image-wrapper {
+                margin: 0;
+
+                &:last-child {
+                    margin: 0;
+                }
+
+                &:first-of-type {
+                    margin: 0 0 50px 0;
+                }
+
+                img {
+                    height: 250px;
+                }
+
+                p {
+                    max-width: 230px;
+                }
+            }
+
+            #arrow-down {
+                display: block;
+                bottom: -63px;
+                left: 84px;
+                height: 45px;
+            }
+        }
+    }
 }
 
 .welcome-page8.variation-3-animation {
-  @media (max-width: 767px) {
-
     .image-container {
-      margin: 35px 0 45px 0;
-    }
-  }
+        margin: 35px auto 70px auto;
+        max-width: 1210px;
 
-  #hero-animation {
-    display: block;
-  }
+        @media (max-width: 767px) {
+            margin: 35px auto 45px auto;
+        }
+
+        .image-wrapper {
+            display: inline-block;
+            vertical-align: top;
+            margin: 0 5% 0 5%;
+            position: relative;
+            width: 25%;
+
+            &:first-of-type {
+                margin: 0 5% 0 0;
+            }
+    
+            &:last-child {
+                margin: 0 0 0 5%;
+            }
+        }
+    }
 }
