+++ media/js/firefox/welcome/welcome8-traffic-cop.js
@@ -1,0 +1,49 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+(function (Mozilla) {
+  'use strict';
+   /* update dataLayer with experiment info */
+    var href = window.location.href;
+
+    var initTrafficCop = function () {
+        if (href.indexOf('v=') !== -1) {
+            if (href.indexOf('v=text') !== -1) {
+                window.dataLayer.push({
+                    'data-ex-variant': 'experiment-hvt-visual-text',
+                    'data-ex-name': 'experiment-hvt-visual'
+                });
+            } else if (href.indexOf('v=image') !== -1) {
+                window.dataLayer.push({
+                    'data-ex-variant': 'experiment-hvt-visual-image',
+                    'data-ex-name': 'experiment-hvt-visual'
+                });
+            } else if (href.indexOf('v=animation') !== -1) {
+                window.dataLayer.push({
+                    'data-ex-variant': 'experiment-hvt-visual-animation',
+                    'data-ex-name': 'experiment-hvt-visual'
+                });
+            } else if (href.indexOf('v=header-text') !== -1) {
+                window.dataLayer.push({
+                    'data-ex-variant': 'experiment-hvt-visual-header-text',
+                    'data-ex-name': 'experiment-hvt-visual'
+                });
+            }
+        } else if (Mozilla.TrafficCop) {
+            var cop = new Mozilla.TrafficCop({
+                id: 'welcome8_experiment_hvt_visual',
+                cookieExpires: 0,
+                variations: {
+                    'v=text': 1,
+                    'v=image': 20,
+                    'v=animation': 20,
+                    'v=header-text': 20,
+                }
+            });
+            cop.init();
+        }
+    };
+    initTrafficCop();
+
+  })(window.Mozilla);
